+++ src/index-dev.ts
@@ -1,14 +1,11 @@
 import { Sustainability } from ".";
 (async () => {
-  const report = await Sustainability.audit(
-    "https://www.homelessentrepreneur.org",
-    {
-      id: "000",
-      connectionSettings: {
-        coldRun: true,
-      },
-    }
-  );
+  const report = await Sustainability.audit("https://www.example.org", {
+    id: "000",
+    connectionSettings: {
+      coldRun: false,
+    },
+  });
 
   console.log(JSON.stringify(report));
 })();